script for deleting account - code style updated

+++ phpfiles/delete_account.php
@@ -1,12 +1,11 @@
 <?php
 session_start();
-if($_SESSION['uid']){
-    if(isset($_POST['delete_account'])){
+if(($_SESSION['uid']) && isset($_POST['delete_account'])){
+    
         include_once ("connection.php");
         $id = $_SESSION['uid'];
 
-
-
+   
             $sql = "DELETE FROM users WHERE id='".$id."'";
             $res = mysqli_query($link, $sql) or die(mysqli_error());
 
@@ -17,9 +16,6 @@ if($_SESSION['uid']){
                 echo "<p>Wystąpił problem ze zmianą loginu.Spróbuj ponownie</p></br><a href='settings.php'>Kliknij by wrócić do ustawień</a>";
             }
 
-    }else{
-        exit();
-    }
 }else{
     header("Location: index.php");
     exit();
